Have maxPasswordLength in XJKeyChain be a UInt32. Rename make_event_ical to show it's applescript nature.

+++ src/Carbon Wrappers/XJKeyChain.m
@@ -76,7 +76,7 @@ static XJKeyChain* defaultKeyChain = nil;
         SecKeychainItemDelete(itemref);
 }
 
-- (void)setMaxPasswordLength:(NSUInteger)length
+- (void)setMaxPasswordLength:(UInt32)length
 {
     if (![self isEqual:defaultKeyChain]) {
         maxPasswordLength = length ;
